+++ packages/drag-in-the-blank/src/index.js
@@ -4,12 +4,10 @@ import {
   ModelSetEvent,
   SessionChangedEvent
 } from '@pie-framework/pie-player-events';
-import { withDragContext } from '@pie-lib/drag';
+import { componentize } from '@pie-lib/mask-markup';
+import Main from './main';
 
-import { DragInTheBlank, componentize } from '@pie-lib/mask-markup';
 // import MarkupAndChoices from './markup-and-choices';
-
-const Root = withDragContext(DragInTheBlank);
 
 // const toConfig = fields => {
 //   return Object.keys(fields).reduce((acc, k) => {
@@ -60,7 +58,7 @@ export default class InlineDropdown extends HTMLElement {
     if (this._model && this._session) {
       const { ids } = componentize(this._model.markup);
       this._session.value = normalize(this._session.value, ids);
-      let elem = React.createElement(Root, {
+      let elem = React.createElement(Main, {
         disabled: this._model.disabled,
         markup: this._model.markup,
         value: this._session.value,