+++ packages/graph-lines/src/main.jsx
@@ -3,7 +3,7 @@ import { GraphLines, lineUtils as utils } from '@pie-lib/charting';
 import PropTypes from 'prop-types';
 import Controls from './controls';
 import CorrectAnswerToggle from '@pie-lib/correct-answer-toggle';
-import { Feedback } from '@pie-lib/render-ui';
+import { Feedback, Collapsible } from '@pie-lib/render-ui';
 import GraphLineControls from './GraphLineControls';
 import { withStyles } from '@material-ui/core/styles';
 
@@ -251,6 +251,16 @@ export class Main extends React.Component {
             onLineClick={this.toggleSelectLine}
           />
         </div>
+        {
+          model.rationale && (
+            <Collapsible
+              labels={{ hidden: 'Show Rationale', visible: 'Hide Rationale' }}
+              className={classes.collapsible}
+            >
+              <div dangerouslySetInnerHTML={{ __html: model.rationale }}/>
+            </Collapsible>
+          )
+        }
         {model.feedback && (
           <Feedback
             correctness={model.correctness.correctness}
@@ -278,6 +288,10 @@ const styles = theme => ({
   },
   toggle: {
     paddingBottom: theme.spacing.unit * 3
+  },
+  collapsible: {
+    paddingTop: theme.spacing.unit * 2,
+    paddingBottom: theme.spacing.unit * 2
   }
 });