+++ packages/math-inline/src/simple-question-block.jsx
@@ -79,19 +79,7 @@ const SimpleQuestionBlock = withStyles(theme => ({
     marginTop: theme.spacing.unit * 2,
     marginBottom: theme.spacing.unit * 2,
     padding: theme.spacing.unit,
-    minHeight: '150px',
-    '& > .mq-math-mode': {
-      '& .mq-non-leaf': {
-        display: 'inline-flex',
-        alignItems: 'center'
-      },
-      '& .mq-non-leaf.mq-fraction': {
-        display: 'inline-block'
-      },
-      '& .mq-paren': {
-        verticalAlign: 'middle'
-      },
-    }
+    minHeight: '150px'
   },
   static: {
     color: 'grey',