+++ packages/passage/demo/config.js
@@ -7,23 +7,31 @@ module.exports = {
       id: '1',
       element: 'passage-element',
       mode: 'view',
-      title: 'Ineskeen Road, July Evening',
-      content: `The bicycles go by in twos and threes -<br/> 
-      There's a dance in Billy Brennan's barn tonight,<br/>
-      And there's the half-talk code of mysteries<br/>
-      And the wink-and-elbow language of delight.<br/>
-      Half-past eight and there is not a spot<br/>
-      Upon a mile of road, no shadow thrown<br/>
-      That might turn out a man or woman, not<br/>
-      A footfall tapping secrecies of stone. <br/>
-      <p/>
-      I have what every poet hates in spite<br/>
-      Of all the solemn talk of contemplation.<br/>
-      Oh, Alexander Selkirk knew the plight<br/>
-      Of being king and government and nation.<br/>
-      A road, a mile of kingdom. I am king<br/>
-      Of banks and stones and every blooming thing.<br/>
-      `
+      passages: [
+        {
+          'title': 'Hope Is the Thing with Feathers',
+          'text': '<p><font style="visibility:hidden">00000</font>Hope is the thing with feathers<br /><font style="visibility:hidden">00000</font>That perches in the soul,<br /><font style="visibility:hidden">00000</font>And sings the tune without the words,<br /><font style="visibility:hidden">00000</font>And never stops at all,</p><p>5<font style="visibility:hidden">0000</font>And sweetest in the gale is heard;<br /><font style="visibility:hidden">00000</font>And sore must be the storm<br /><font style="visibility:hidden">00000</font>That could abash the little bird<br /><font style="visibility:hidden">00000</font>That kept so many warm.</p><p><font style="visibility:hidden">00000</font>I&#8217;ve heard it in the chillest land,<br />10<font style="visibility:hidden">000</font>And on the strangest sea;<br /><font style="visibility:hidden">00000</font>Yet, never, in extremity,<br /><font style="visibility:hidden">00000</font>It asked a crumb of me.</p><p><span class="no-number">&#34;Hope is the Thing With Feathers&#34; by Emily Dickinson, from <em>The Complete Poems of Emily Dickinson</em>. In the public domain.</span></p>'
+        },
+        {
+          title: 'Ineskeen Road, July Evening',
+          text: `The bicycles go by in twos and threes -<br/> 
+          There's a dance in Billy Brennan's barn tonight,<br/>
+          And there's the half-talk code of mysteries<br/>
+          And the wink-and-elbow language of delight.<br/>
+          Half-past eight and there is not a spot<br/>
+          Upon a mile of road, no shadow thrown<br/>
+          That might turn out a man or woman, not<br/>
+          A footfall tapping secrecies of stone. <br/>
+          <p/>
+          I have what every poet hates in spite<br/>
+          Of all the solemn talk of contemplation.<br/>
+          Oh, Alexander Selkirk knew the plight<br/>
+          Of being king and government and nation.<br/>
+          A road, a mile of kingdom. I am king<br/>
+          Of banks and stones and every blooming thing.<br/>
+          `
+        }
+      ]
     }
   ]
 };