+++ packages/demo/CHANGELOG.md
@@ -3,6 +3,118 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.19](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.18...@pie-ui/demo@0.3.19) (2019-05-21)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.18](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.17...@pie-ui/demo@0.3.18) (2019-05-16)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.17](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.16...@pie-ui/demo@0.3.17) (2019-05-15)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.16](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.15...@pie-ui/demo@0.3.16) (2019-05-14)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.15](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.14...@pie-ui/demo@0.3.15) (2019-05-14)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.14](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.13...@pie-ui/demo@0.3.14) (2019-05-13)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.13](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.12...@pie-ui/demo@0.3.13) (2019-05-10)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.12](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.11...@pie-ui/demo@0.3.12) (2019-05-10)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.11](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.10...@pie-ui/demo@0.3.11) (2019-05-09)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.10](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.9...@pie-ui/demo@0.3.10) (2019-05-08)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.9](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.8...@pie-ui/demo@0.3.9) (2019-05-08)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.8](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.7...@pie-ui/demo@0.3.8) (2019-05-07)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.7](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.6...@pie-ui/demo@0.3.7) (2019-05-03)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
+## [0.3.6](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.5...@pie-ui/demo@0.3.6) (2019-05-03)
+
+**Note:** Version bump only for package @pie-ui/demo
+
+
+
+
+
 ## [0.3.5](https://github.com/pie-framework/pie-ui/compare/@pie-ui/demo@0.3.4...@pie-ui/demo@0.3.5) (2019-05-02)
 
 **Note:** Version bump only for package @pie-ui/demo