+++ packages/hotspot/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-ui/hotspot@1.1.0...@pie-ui/hotspot@1.1.1) (2019-05-16)
+
+**Note:** Version bump only for package @pie-ui/hotspot
+
+
+
+
+
+# [1.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-ui/hotspot@1.0.4...@pie-ui/hotspot@1.1.0) (2019-05-15)
+
+
+### Bug Fixes
+
+* Updated pie-lib/render-ui version (to use className property). Added snapshot test for rationale. ([19ab3f9](https://github.com/pie-framework/pie-elements/commit/19ab3f9))
+
+
+### Features
+
+* Added rationale support. ([47e2503](https://github.com/pie-framework/pie-elements/commit/47e2503))
+
+
+
+
+
+## [1.0.4](https://github.com/pie-framework/pie-elements/compare/@pie-ui/hotspot@1.0.3...@pie-ui/hotspot@1.0.4) (2019-05-08)
+
+**Note:** Version bump only for package @pie-ui/hotspot
+
+
+
+
+
+## [1.0.3](https://github.com/pie-framework/pie-elements/compare/@pie-ui/hotspot@1.0.2...@pie-ui/hotspot@1.0.3) (2019-05-07)
+
+**Note:** Version bump only for package @pie-ui/hotspot
+
+
+
+
+
 ## [1.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-ui/hotspot@1.0.1...@pie-ui/hotspot@1.0.2) (2019-05-02)
 
 **Note:** Version bump only for package @pie-ui/hotspot