+++ packages/inline-dropdown/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.1.3...@pie-ui/inline-dropdown@1.2.0) (2019-05-21)
+
+
+### Features
+
+* **mask-markup-ui:** added math rendering and prompts to all elements ([94b7b97](https://github.com/pie-framework/pie-ui/commit/94b7b97))
+
+
+
+
+
+## [1.1.3](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.1.2...@pie-ui/inline-dropdown@1.1.3) (2019-05-14)
+
+**Note:** Version bump only for package @pie-ui/inline-dropdown
+
+
+
+
+
+## [1.1.2](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.1.1...@pie-ui/inline-dropdown@1.1.2) (2019-05-13)
+
+**Note:** Version bump only for package @pie-ui/inline-dropdown
+
+
+
+
+
+## [1.1.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.1.0...@pie-ui/inline-dropdown@1.1.1) (2019-05-10)
+
+**Note:** Version bump only for package @pie-ui/inline-dropdown
+
+
+
+
+
+# [1.1.0](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.0.4...@pie-ui/inline-dropdown@1.1.0) (2019-05-10)
+
+
+### Features
+
+* Added prompt support. ([1e8bebc](https://github.com/pie-framework/pie-ui/commit/1e8bebc))
+
+
+
+
+
+## [1.0.4](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.0.3...@pie-ui/inline-dropdown@1.0.4) (2019-05-08)
+
+**Note:** Version bump only for package @pie-ui/inline-dropdown
+
+
+
+
+
+## [1.0.3](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.0.2...@pie-ui/inline-dropdown@1.0.3) (2019-05-03)
+
+**Note:** Version bump only for package @pie-ui/inline-dropdown
+
+
+
+
+
+## [1.0.2](https://github.com/pie-framework/pie-ui/compare/@pie-ui/inline-dropdown@1.0.1...@pie-ui/inline-dropdown@1.0.2) (2019-05-03)
+
+
+### Bug Fixes
+
+* **dependencies:** updated version for the mask-markup elements ([5ba17e2](https://github.com/pie-framework/pie-ui/commit/5ba17e2))
+
+
+
+
+
 ## 1.0.1 (2019-05-02)