+++ packages/passage/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.2.3...@pie-ui/passage@1.2.4) (2019-05-16)
+
+**Note:** Version bump only for package @pie-ui/passage
+
+
+
+
+
+## [1.2.3](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.2.2...@pie-ui/passage@1.2.3) (2019-05-15)
+
+
+### Bug Fixes
+
+* don't render title if only one passage ([#104](https://github.com/pie-framework/pie-ui/issues/104)) ([5632ef9](https://github.com/pie-framework/pie-ui/commit/5632ef9))
+
+
+
+
+
+## [1.2.2](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.2.1...@pie-ui/passage@1.2.2) (2019-05-14)
+
+
+### Bug Fixes
+
+* update passage title style ([f8c779f](https://github.com/pie-framework/pie-ui/commit/f8c779f))
+
+
+
+
+
+## [1.2.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.2.0...@pie-ui/passage@1.2.1) (2019-05-14)
+
+**Note:** Version bump only for package @pie-ui/passage
+
+
+
+
+
+# [1.2.0](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.1.1...@pie-ui/passage@1.2.0) (2019-05-13)
+
+
+### Features
+
+* initial paired passages support ([c978fa1](https://github.com/pie-framework/pie-ui/commit/c978fa1))
+
+
+
+
+
 ## [1.1.1](https://github.com/pie-framework/pie-ui/compare/@pie-ui/passage@1.1.0...@pie-ui/passage@1.1.1) (2019-05-01)
 
 **Note:** Version bump only for package @pie-ui/passage